--- core/controller/metatag/models.go
@@ -1,7 +1,0 @@
-package metatag
-
-import "github.com/horizoncd/horizon/pkg/metatag/models"
-
-type CreateMetatagsRequest struct {
-	Metatags []*models.Metatag
-}
