+++ db/20230512.sql
@@ -604,11 +604,11 @@ CREATE TABLE `tb_webhook_log`
 -- metatag table
 CREATE TABLE `tb_metatag`
 (
-    `tag_key`            varchar(64)  NOT NULL DEFAULT '' comment 'key of the metatag',
-    `tag_value`          varchar(128) NOT NULL DEFAULT '' comment 'value of the metatag',
-    `tag_value_identity` varchar(64)  NOT NULL DEFAULT '' comment 'identity of value of the metatag',
-    `created_at`         datetime     NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    `updated_at`         datetime     NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    `tag_key`     varchar(64)  NOT NULL DEFAULT '' comment 'key of the metatag',
+    `tag_value`   varchar(128) NOT NULL DEFAULT '' comment 'value of the metatag',
+    `description` varchar(64)  NOT NULL DEFAULT '' comment 'description',
+    `created_at`  datetime     NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `updated_at`  datetime     NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     UNIQUE KEY `idx_key_value` (`tag_key`, `tag_value`)
 ) ENGINE = InnoDB
   AUTO_INCREMENT = 1
